requirements.txt
# ---- Core Backend ----
fastapi
uvicorn[standard]
pydantic
python-dotenv
loguru

# ---- LLM & RAG ----
langchain
langchain-community
langchain-openai
llama-index
llama-index-vector-stores-pinecone
langchain-text-splitters


# ---- Vector DB ----
pinecone-client>=3.0.0


# ---- PDF Processing ----
pypdf
pdfplumber

# ---- Embeddings & Tokenization ----
tiktoken
sentence-transformers

# ---- Hybrid Search ----
rank-bm25
scikit-learn
numpy

# ---- Re-ranking (Optional but recommended) ----
torch
transformers

# ---- Utilities ----
requests
tenacity

# ---- Testing ----
pytest
